+++ src/lib/components/Card/index.js
@@ -1,4 +1,4 @@
-import { Card } from 'antd'
+import Card from './Base'
 import CardTable from './Table'
 
 Card.Table = CardTable